Modularize GFP epoch cleaning as plugin
Former-commit-id: 531d18c89c03602339648bdc4a8f6e48e34dbd99
