+++ docs/mdx/wavelet-thresholding.mdx
@@ -53,7 +53,7 @@ These unit tests exercise the new hard-thresholding branch and the ERP-aware fil
 
 ## PDF Reporting
 
-`wavelet_threshold` and `generate_wavelet_report` now live in the same module. You can drop `src/autoclean/functions/preprocessing/wavelet_thresholding.py` into a lightweight environment and retain both cleaning and reporting.
+`wavelet_threshold` and `generate_wavelet_report` now live beside the mixin inside `src/autoclean/mixins/signal_processing/wavelet_threshold/`. Drop that folder into a lightweight environment to retain both cleaning and reporting.
 
 Use the helper script to create a PDF comparison of pre/post wavelet cleaning:
 
@@ -66,7 +66,7 @@ python scripts/wavelet_report.py /path/to/TD1v2_ica_raw.set \
 The report includes summary statistics, a top-channel waveform overlay, and a leaderboard of channels ranked by peak-to-peak reduction. Adjust `--wavelet`, `--level`, or `--picks all` to explore different configurations. The PDF now also charts the average power spectral density pre/post cleaning; pass `--psd-csv` to export band power reductions (delta–gamma) as a CSV. To call the API directly, import from the unified module:
 
 ```python
-from autoclean.functions.preprocessing.wavelet_thresholding import generate_wavelet_report
+from autoclean.mixins.signal_processing.wavelet_threshold.processing import generate_wavelet_report
 
 report = generate_wavelet_report(raw, "reports/example_wavelet.pdf")
 ```
