+++ scripts/wavelet_report.py
@@ -17,9 +17,10 @@ def _load_generate_wavelet_report():
         Path(__file__).resolve().parents[1]
         / "src"
         / "autoclean"
-        / "functions"
-        / "preprocessing"
-        / "wavelet_thresholding.py"
+        / "mixins"
+        / "signal_processing"
+        / "wavelet_threshold"
+        / "processing.py"
     )
     spec = importlib.util.spec_from_file_location(
         "autoclean.reporting.wavelet_report_cli",
