+++ src/autoclean/mixins/signal_processing/autoreject/manifest.json
@@ -1,0 +1,30 @@
+{
+  "id": "autoclean.autoreject",
+  "name": "AutoReject Cleaning",
+  "version": "1.0.0",
+  "authors": [
+    {
+      "name": "Ernest Pedapati",
+      "email": "ernest.pedapati@cchmc.org"
+    }
+  ],
+  "summary": "AutoReject wrapper that accepts MNE Raw or Epochs data and returns cleaned epochs with the reject log.",
+  "entry_points": [
+    {
+      "function": "autoclean.mixins.signal_processing.autoreject.processing.run_autoreject",
+      "input": "mne.Epochs",
+      "output": "AutorejectResult"
+    },
+    {
+      "function": "autoclean.mixins.signal_processing.autoreject.processing.run_autoreject_raw",
+      "input": "mne.io.Raw",
+      "output": "AutorejectResult"
+    }
+  ],
+  "mixins": [
+    {
+      "class": "autoclean.mixins.signal_processing.autoreject.AutorejectMixin",
+      "method": "apply_autoreject"
+    }
+  ]
+}
