+++ src/autoclean/mixins/signal_processing/star_cleaning/manifest.json
@@ -1,0 +1,25 @@
+{
+  "id": "autoclean.star_cleaning",
+  "name": "STAR Spatial Cleaning",
+  "version": "1.0.0",
+  "authors": [
+    {
+      "name": "Ernest Pedapati",
+      "email": "ernest.pedapati@cchmc.org"
+    }
+  ],
+  "summary": "Spatial filtering using meegkit's STAR algorithm with Raw in/out parity.",
+  "entry_points": [
+    {
+      "function": "autoclean.mixins.signal_processing.star_cleaning.processing.run_star_cleaning",
+      "input": "mne.io.Raw",
+      "output": "mne.io.Raw"
+    }
+  ],
+  "mixins": [
+    {
+      "class": "autoclean.mixins.signal_processing.star_cleaning.StarCleaningMixin",
+      "method": "apply_star_cleaning"
+    }
+  ]
+}
